requirements.txt
AccessControl==5.0
Acquisition==4.7
AuthEncoding==4.2
beautifulsoup4==4.9.3
BTrees==4.7.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
Chameleon==3.8.1
chardet==3.0.4
DateTime==4.3
DocumentTemplate==4.0
ExtensionClass==4.5.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
MultiMapping==4.1
oauthlib==3.1.0
PasteDeploy==2.1.1
Persistence==3.0
persistent==4.6.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-gettext==4.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
RestrictedPython==5.1
roman==3.3
rsa==4.6
six==1.15.0
soupsieve==2.0.1
transaction==3.0.0
uritemplate==3.0.1
urllib3==1.26.2
waitress==1.4.4
WebOb==1.8.6
WebTest==2.0.35
WSGIProxy2==0.4.6
z3c.pt==3.3.0
zc.lockfile==2.0
ZConfig==3.5.0
zExceptions==4.1
ZODB==5.6.0
zodbpickle==2.0.0
zope.annotation==4.7.0
zope.browser==2.3
zope.browsermenu==4.4
zope.browserpage==4.4.0
zope.browserresource==4.4
zope.cachedescriptors==4.3.1
zope.component==4.6.2
zope.configuration==4.4.0
zope.container==4.4.0
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.dottedname==4.3
zope.event==4.5.0
zope.exceptions==4.4
zope.filerepresentation==5.0.0
zope.globalrequest==1.5
zope.hookable==5.0.1
zope.i18n==4.7.0
zope.i18nmessageid==5.0.1
zope.interface==5.2.0
zope.lifecycleevent==4.3
zope.location==4.2
zope.pagetemplate==4.5.0
zope.processlifetime==2.3.0
zope.proxy==4.3.5
zope.ptresource==4.2.0
zope.publisher==5.2.1
zope.schema==6.0.0
zope.security==5.1.1
zope.sequencesort==4.1.2
zope.site==4.4.0
zope.size==4.3
zope.structuredtext==4.3
zope.tal==4.4
zope.tales==5.1
zope.testbrowser==5.5.1
zope.testing==4.7
zope.traversing==4.4.1
zope.viewlet==4.2.1
